old-tests/submission/W3C/domhtml/HTMLBodyElement06.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-body-vlink" rel="help">
<link rel="author" title="NIST">
</head>
<body vlink="#00ffff"> 
<p>Hello, World</p>


<script>
var domtest = async_test("The vLink IDL attribute of the body element must reflect the element's vlink content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vvlink;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("body");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vvlink = testNode.vLink;

      assert_equals(vvlink,"#00ffff");
       
    domtest.done();
}));

</script>
<p>The vLink IDL attribute of the body element must reflect the element's vlink content attribute.</p>
<div id="log"></div>
</body></html>
